feat: use locator
